fix: modified url for local storage provider

+++ src/features/upload-file/factories/delete-files-factory.ts
@@ -1,4 +1,9 @@
-import { BulkActionResponse, After, ActionRequest, ActionContext } from 'adminjs/types/src'
+import {
+  ActionContext,
+  ActionRequest,
+  After,
+  BulkActionResponse,
+} from 'adminjs/types/src'
 import { BaseProvider } from '../providers'
 import { UploadOptionsWithDefault } from '../types/upload-options.type'
 import { deleteFile } from '../utils/delete-file'
@@ -12,10 +17,13 @@ export const deleteFilesFactory = (
   context: ActionContext,
 ): Promise<BulkActionResponse> {
   const { records = [] } = context
-
-  await Promise.all(records.map(async (record) => {
-    await deleteFile(uploadOptionsWithDefault, provider, context, record)
-  }))
+  if (request.method === 'post') {
+    await Promise.all(
+      records.map(async (record) => {
+        await deleteFile(uploadOptionsWithDefault, provider, context, record)
+      }),
+    )
+  }
 
   return response
 }
